backend-main/backend-main/requirements.txt
amqp
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
billiard==4.2.0
celery==5.3.6
newrelic
pika
cloudinary==1.36.0

dj-database-url==2.0.0
dj-rest-auth==2.2.7
dj-static==0.0.6
Django==4.2.4
django-allauth==0.52.0

django-cors-headers==4.2.0

django-crontab-0.7.1
django-email-verification==0.3.1
django-filter==23.1
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.1.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
gcloud==0.18.3
googleapis-common-protos==1.60.0
google_auth_oauthlib
gunicorn==21.2.0

oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
phonenumberslite==8.13.20
prometheus_client==0.19.0
prompt_toolkit==3.0.43
protobuf==4.24.3
# psycopg2==2.9.6
# psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==4.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
SQLAlchemy==2.0.19
sqlparse==0.4.4
django_elasticsearch_dsl
setuptools
psycopg2-binary

static3==0.7.0
swagger-ui-bundle==0.0.9
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.5.0
algoliasearch

django_redis
moesifdjango
pandas 
openpyxl
algoliasearch-django
django-csp
django-permissions-policy
django-sonar
aes-everywhere

google-auth-oauthlib
